+++ AzureActiveDirectory/OAuth2/Servlet/docs/sample-03-multiple-resource-server.md
@@ -118,7 +118,7 @@ public class ClientApplication {
 
 #### 2.1.2.2. WebSecurityConfiguration.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.client.configuration;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.client.configuration;
 
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -142,7 +142,7 @@ public class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
 
 #### 2.1.2.3. ApplicationConfiguration.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.client.configuration;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.client.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -169,7 +169,7 @@ public class ApplicationConfiguration {
 
 #### 2.1.2.4. HomeController.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.client.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.client.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -186,7 +186,7 @@ public class HomeController {
 
 #### 2.1.2.5. ResourceServer1Controller.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.client.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.client.controller;
 
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
 import org.springframework.security.oauth2.client.annotation.RegisteredOAuth2AuthorizedClient;
@@ -220,7 +220,7 @@ public class ResourceServer1Controller {
 
 #### 2.1.2.6. ResourceServer2Controller.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.client.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.client.controller;
 
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
 import org.springframework.security.oauth2.client.annotation.RegisteredOAuth2AuthorizedClient;
@@ -254,7 +254,7 @@ public class ResourceServer2Controller {
 
 #### 2.1.2.7. ResourceServerAllController.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.client.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.client.controller;
 
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
 import org.springframework.security.oauth2.client.annotation.RegisteredOAuth2AuthorizedClient;
@@ -352,7 +352,7 @@ spring:
 
 ### 2.2.2.1. ResourceServerApplication.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server1;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server1;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -368,7 +368,7 @@ public class ResourceServerApplication {
 
 #### 2.2.2.2. WebSecurityConfiguration.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server1.configuration;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server1.configuration;
 
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -393,7 +393,7 @@ public class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
 
 #### 2.2.2.3. ApplicationConfiguration.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server1.configuration;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server1.configuration;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.security.oauth2.resource.OAuth2ResourceServerProperties;
@@ -465,7 +465,7 @@ public class ApplicationConfiguration {
 
 #### 2.2.2.4. HomeController.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server1.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server1.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -538,7 +538,7 @@ spring:
 
 #### 2.3.2.1. ResourceServerApplication.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server2;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server2;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -554,7 +554,7 @@ public class ResourceServerApplication {
 
 #### 2.3.2.2. WebSecurityConfiguration.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server2.configuration;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server2.configuration;
 
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -579,7 +579,7 @@ public class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
 
 #### 2.3.2.3. ApplicationConfiguration.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server2.configuration;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server2.configuration;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.security.oauth2.resource.OAuth2ResourceServerProperties;
@@ -651,7 +651,7 @@ public class ApplicationConfiguration {
 
 #### 2.3.2.4. HomeController.java
 ```java
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server2.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server2.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
