--- AzureActiveDirectory/OAuth2/Servlet/sample-03-multiple-resource-server/sample-03-client/src/main/java/com/azure/spring/sample/active/directory/oauth2/sample03/client/ClientApplication.java
@@ -1,12 +1,0 @@
-package com.azure.spring.sample.active.directory.oauth2.sample03.client;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ClientApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ClientApplication.class, args);
-    }
-}
