+++ AzureActiveDirectory/OAuth2/Servlet/sample-03-multiple-resource-server/sample-03-client/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample03/client/controller/ResourceServer1Controller.java
@@ -1,4 +1,4 @@
-package com.azure.spring.sample.active.directory.oauth2.sample03.client.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.client.controller;
 
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
 import org.springframework.security.oauth2.client.annotation.RegisteredOAuth2AuthorizedClient;
@@ -9,20 +9,20 @@ import org.springframework.web.reactive.function.client.WebClient;
 import static org.springframework.security.oauth2.client.web.reactive.function.client.ServletOAuth2AuthorizedClientExchangeFilterFunction.oauth2AuthorizedClient;
 
 @RestController
-public class ResourceServer2Controller {
+public class ResourceServer1Controller {
 
     private final WebClient webClient;
 
-    public ResourceServer2Controller(WebClient webClient) {
+    public ResourceServer1Controller(WebClient webClient) {
         this.webClient = webClient;
     }
 
-    @GetMapping("/resource-server-2")
-    public String hello(@RegisteredOAuth2AuthorizedClient("client-1-resource-server-2") OAuth2AuthorizedClient client1ResourceServer2) {
+    @GetMapping("/resource-server-1")
+    public String hello(@RegisteredOAuth2AuthorizedClient("client-1-resource-server-1") OAuth2AuthorizedClient client1ResourceServer1) {
         return webClient
             .get()
-            .uri("http://localhost:8082")
-            .attributes(oauth2AuthorizedClient(client1ResourceServer2))
+            .uri("http://localhost:8081")
+            .attributes(oauth2AuthorizedClient(client1ResourceServer1))
             .retrieve()
             .bodyToMono(String.class)
             .block();
