+++ AzureActiveDirectory/OAuth2/Servlet/sample-03-multiple-resource-server/sample-03-resource-server-1/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample03/resource/server1/configuration/ApplicationConfiguration.java
@@ -1,4 +1,4 @@
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server1.configuration;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server1.configuration;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.security.oauth2.resource.OAuth2ResourceServerProperties;
