+++ AzureActiveDirectory/OAuth2/Servlet/sample-03-multiple-resource-server/sample-03-resource-server-1/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample03/resource/server1/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.azure.spring.sample.active.directory.oauth2.sample03.client.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server1.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -8,6 +8,6 @@ public class HomeController {
 
     @GetMapping("/")
     public String home() {
-        return "Hello, this is sample-03-client.";
+        return "Hello, this is resource-server-1.";
     }
 }
