+++ AzureActiveDirectory/OAuth2/Servlet/sample-03-multiple-resource-server/sample-03-resource-server-2/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample03/resource/server2/configuration/WebSecurityConfiguration.java
@@ -1,4 +1,4 @@
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server1.configuration;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server2.configuration;
 
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
