+++ AzureActiveDirectory/OAuth2/Servlet/sample-03-multiple-resource-server/sample-03-resource-server-2/src/main/java/com/azure/spring/sample/active/directory/oauth2/servlet/sample03/resource/server2/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.azure.spring.sample.active.directory.oauth2.sample03.resource.server2.controller;
+package com.azure.spring.sample.active.directory.oauth2.servlet.sample03.resource.server2.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
